% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ace_multi.R
\name{ace_multi}
\alias{ace_multi}
\title{Ancestral Character Estimation for Multiple Traits}
\usage{
ace_multi(x, phy)
}
\arguments{
\item{x}{A species \code{×} k matrix, where the column sums equal 1.
These values represent the known states or contributions of each
species to each cluster/biome to be represented in the tree.}

\item{phy}{A dated and rooted phylogenetic tree of the class "phylo".
The root is assumed to be the unique node with no incoming edge.}
}
\value{
A list with the following elements:
\itemize{
  \item \code{ancestral_states} The estimates of the ancestral
  character values.
  \item \code{CI95} The estimated 95 percent confidence intervals.
  \item \code{se} The standard-errors of estimated rates.
  \item \code{tree} The pruned tree with tips matching the input
  matrix \code{x} above.
  \item \code{piecolors} List of HUE colors for plotting nodepies in
  the tree.
}
}
\description{
This function reconstructs ancestral states for multiple
continuous (numeric) trait using phylogenetic independent
contrasts (PIC; Felsenstein, 1985).
}
\examples{
\donttest{
old.par <- par(no.readonly = TRUE)
require(ape)
data(africa)
m <- ace_multi(africa$theta, africa$phylo)

par(mai=rep(0,4))
par(mfrow=c(2,1))
plot(m$tree, type = "fan", show.tip.label=FALSE,
     open.angle = 180, edge.width=0.5)

nodelabels(pie = m$ancestral_states, cex = 0.23,
           piecol=m$piecolors, lwd=0.0001)

plot_structure(africa$omega, shp = africa$polys, legend = TRUE)
par(old.par)
}

}
\references{
Felsenstein, J. (1985) Phylogenies and the comparative
method. \emph{American Naturalist} \strong{125}: 1-15.
}
