% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community.R
\name{as.community}
\alias{as.community}
\alias{read.community}
\alias{as.community.matrix}
\alias{print.community}
\alias{as.community.matrix}
\alias{as.community.data.frame}
\alias{read.community}
\title{Sparse community matrices}
\usage{
as.community(x, ...)

as.community.matrix(x)

as.community.data.frame(x)

read.community(file, ...)
}
\arguments{
\item{splist}{Species list}

\item{tree}{a phylogenetic tree (object of class phylo)}

\item{nodes}{node list}

\item{output.splist}{return species list}
}
\description{
read.community reads in file containing occurence data and returns a sparse
matrix.
}
\details{
If species is not in the tip label species is added to the genus or family
level when possible
}
\examples{
tree <- read.tree(text ="((t1:1,t2:1)N2:1,(t3:1,t4:1)N3:1)N1;")
com <- matrix(c(1,0,1,1,0,0,
                1,0,0,1,1,0,
                1,1,1,1,1,1,
                0,0,1,1,0,1), 6, 4,
              dimnames=list(paste0("g",1:6), tree$tip.label))
com <- as.community(com)
com

}
\keyword{cluster}
