% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_matrix.R
\name{av.matrix}
\alias{av.matrix}
\title{Mean matrix from a set of pairwise distance matrices}
\usage{
av.matrix(files, tips, verbose = TRUE, ...)
}
\arguments{
\item{files}{list of pairwise distance matrices stored as CSVs or .rds
with the same dimensions.}

\item{tips}{list of site or grid names}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
average pairwise distance matrix
}
\description{
This function generates the mean pairwise distance matrix from a set
many pairwise distance matrices. Note: all matrices should be of the same dimension.
}
\examples{
rm(list = ls())
wd <- tempdir()
setwd(wd)

## Generate the sets of pairwise distance matrices
require(ape)
set.seed(1)
tree <- read.tree(text ="((t1:1,t2:1)N2:1,(t3:1,t4:1)N3:1)N1;")
tree <- rmtree(10, 4)
com <- matrix(c(1,0,1,1,0,0,
               1,0,0,1,1,0,
                1,1,1,1,1,1,
                0,0,1,1,0,1), 6, 4,
              dimnames=list(paste0("g",1:6), paste0("t", 1:4)))


for (i in 1:length(tree)) {
  pbc <- phylobeta(com, tree[[i]])
  write.csv(pbc, file = xzfile(paste0("matrix", i, ".xz")))
}

files <- list.files(wd, pattern="*.xz", full.names=TRUE)
tr <- colnames(read.csv(files[1], row.names = 1, stringsAsFactors = FALSE))
tr1 <- sort(tr)
av.matrix(files, tr1)
}
