% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/background.R
\name{backg}
\alias{backg}
\alias{randpoints}
\title{Sample background points}
\usage{
backg(calib, spatkde, size = 10000, ...)

randpoints(ras, size, prob = NULL, ...)
}
\arguments{
\item{calib}{A SpatRaster of the species calibration area. If a polygon, 
convert to SpatRaster using rasterize.}

\item{spatkde}{A weighted or unweighted Gaussian Kernel Density estimate 
(KDE) for all input occurrence records.}

\item{size}{An positive integer of the number of samples to generate.}

\item{...}{further arguments passed to or from other methods.}

\item{ras}{Input SpatRaster}

\item{prob}{Vector of probability weights for obtaining the points sampled.}
}
\value{
A dataframe containing the generated background points.

A dataframe with sampled points.
}
\description{
Generates background sample as null for species distribution 
modeling and other things.

Sample random background points using a vector of probability 
weights.
}
