% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo_beta.R
\name{beta.core}
\alias{beta.core}
\title{Taxonomic (non-phylogenetic) beta diversity}
\usage{
beta.core(x, index.family = "sorensen", index = "Simpson")
}
\description{
Taxonomic (non-phylogenetic) beta diversity
}
