% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo_beta.R
\name{beta_core}
\alias{beta_core}
\alias{beta_diss}
\title{Taxonomic (non-phylogenetic) beta diversity}
\usage{
beta_core(x)

beta_diss(x, index.family = "sorensen")
}
\arguments{
\item{x}{an object of class Matrix or matrix.}

\item{index.family}{family of dissimilarity indices, partial match of
"sorensen" or "jaccard".}
}
\description{
Taxonomic (non-phylogenetic) beta diversity
}
\seealso{
\code{\link{betapart.core}}
}
