% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.algorithm.R
\name{clustering.algorithm}
\alias{clustering.algorithm}
\title{Cluster algorithm selection and validation}
\usage{
clustering.algorithm(
  x,
  plot = FALSE,
  rot_angle = 45,
  cex = 0.7,
  col = "grey",
  ...
)
}
\arguments{
\item{x}{a numeric matrix, data frame or "dist" object.}

\item{plot}{logical if results should be plotted.}

\item{rot_angle}{rotation angle of the barplot x-axis labels.}

\item{cex}{character (or symbol) expansion: a numerical vector for
the size of the x-axis labels}

\item{col}{color of the plot.}

\item{\dots}{arguments passed among methods.}
}
\value{
\itemize{
  \item A numeric value corresponding to the good clustering algorithm
  for the distance matrix
  \item If plot = TRUE, a barplot of cophenetic correlation for all
  the clustering algorithms is drawn.
}
}
\description{
This function contrasts different hierarchical clustering algorithms
on the phylogenetic beta diversity matrix for degree of
data distortion using Sokal & Rohlf’s (1962) cophenetic
correlation coefficient.
}
\examples{
require(betapart)
data(africa)
tree <- africa$phylo
M <- sampl2sparse(africa$comm, method = "nonphylo")

bc <- beta_diss(M)
clustering.algorithm(bc, plot = TRUE)
}
\references{
Sokal, R.R. & Rohlf, F.J. (1962) The comparison of dendrograms by
objective methods. \emph{Taxon} \strong{11}: 33–40.
}
\keyword{bioregion}
