% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/com2sparse.R
\name{comm2sparse}
\alias{comm2sparse}
\title{Convert community matrix to sparse community matrix}
\usage{
comm2sparse(x)
}
\arguments{
\item{x}{a community matrix or data frame}
}
\description{
Converts a species by sites community matrix to a saprse community matrix.
Expects a matrix with rows as sites and columns as species or taxa.
}
\note{
See \url{https://github.com/darunabas/phyloregion} for more
details and tutorials.
}
\examples{
require(sp)
com <- matrix(c(1,0,1,1,0,0,
                1,0,0,1,1,0,
                1,1,1,1,1,1,
                0,0,1,1,0,1), 6, 4,
              dimnames=list(paste0("g",1:6), paste0("sp", 1:4)))
comm2sparse(com)
}
\keyword{Conversion}
