% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/darwinize.R
\name{darwinize}
\alias{darwinize}
\title{Darwin's naturalization conundrum}
\usage{
darwinize(x, phy, iter)
}
\arguments{
\item{x}{A data frame or vector containing a species list}

\item{phy}{A phylogenetic tree}

\item{iter}{Numeric, the number of permutations}
}
\value{
a vector with the mean distance between species
}
\description{
\code{darwinize} computes standard effective size of
mean phylogenetic distance between nearest neighbors.
}
\examples{
data(africa)
tree <- africa$phylo
y <- sample(tree$tip.label, 10, replace = TRUE)

darwinize(y, tree, 100)

}
\author{
Barnabas H. Daru \email{darunabas@gmail.com}
}
\keyword{bioregion}
