% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/background.R
\name{dirpath}
\alias{dirpath}
\title{Get current directory}
\usage{
dirpath(path)
}
\arguments{
\item{path}{Character vector of the directory path names.}
}
\value{
A character vector containing the name of the current directory.
}
\description{
Gets the path of the current directory.
}
