% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evol_distinct.R
\name{evol_distinct}
\alias{evol_distinct}
\title{Species' evolutionary distinctiveness}
\usage{
evol_distinct(
  tree,
  type = c("equal.splits", "fair.proportion"),
  scale = FALSE,
  use.branch.lengths = TRUE
)
}
\arguments{
\item{tree}{an object of class \code{phylo}.}

\item{type}{a) equal splits (Redding and Mooers 2006) or b) fair proportions
(Isaac et al., 2007)}

\item{scale}{The scale option refers to whether or not the phylogeny should
be scaled to a depth of 1 or, in the case of an ultrametric tree, scaled such
that branch lengths are relative.}

\item{use.branch.lengths}{If use.branch.lengths=FALSE, then all branch
lengths are changed to 1.}
}
\value{
a named vector with species scores.
}
\description{
Calculates evolutionary distinctiveness measures for a suite of species by:
a) equal splits (Redding and Mooers 2006)
b) fair proportions (Isaac et al., 2007).
This is an reimplemantation of the picante function allowing multifurcations
and can be orders of magnitude faster.
}
\seealso{
\code{\link[picante]{evol.distinct}}
}
\author{
Klaus Schliep based on version in picante
}
