% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_traits.R
\name{get_trait}
\alias{get_trait}
\title{Plant traits from online databases}
\usage{
get_trait(X, database = "FNA")
}
\arguments{
\item{X}{A vector of species names}

\item{database}{The database to search for traits. Available databases include:
\itemize{
  \item \dQuote{eflora_US} = Flora of North America.
  \item \dQuote{FNA} = Beta version of Flora of North America.
  \item \dQuote{plantzafrica} = Plants of southern Africa.
  \item \dQuote{wikipedia} = Wikipedia free online encyclopedia.
  \item \dQuote{reflora} = Flora of Brazil
}}
}
\value{
A dataframe containing the traits data for the species of interest.
}
\description{
Get plant height data from online databases.
This function requires Internet connection.
If the desired database uses java script to store the data
(e.g. Flora of Brazil),
the function requires phantomjs (.exe for windows) stored in
the workind directory.
}
\examples{
sp <- c("Graptopetalum bartramii", "Salix raupii",
       "Berberis harrisoniana", "Quercus arkansana", "Aloe ferox")
get_trait(sp, database = "FNA")
}
