% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_Clusters.R
\name{getClusters}
\alias{getClusters}
\title{Get descendant nodes of phylogeny at a given time depth.}
\usage{
getClusters(tree, cut = 2, ...)
}
\arguments{
\item{tree}{is a dated phylogenetic tree with branch lengths stored
as a phylo object (as in the \code{ape} package).}

\item{cut}{the slice time}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
A list of descendants
}
\description{
\code{getClusters} returns the tips that descend from a given node or time depth on a dated phylogenetic tree.
}
\examples{
require(ape)
data(bird.orders)
plot(bird.orders)
axisPhylo(side = 1)
abline(v=23)
getClusters(bird.orders, 23)
}
\references{
Schliep, K.P. (2010) phangorn: phylogenetic analysis in
R. \emph{Bioinformatics} \strong{27}: 592–593.
}
\keyword{bioregion}
