% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_phy_age.R
\name{getNage}
\alias{getNage}
\title{Get age of nodes on phylogenetic tree}
\usage{
getNage(phy, n)
}
\arguments{
\item{phy}{A phylogenetic tree of the class \dQuote{phylo}.}

\item{n}{An index or indices indicating the node positions.}
}
\value{
A numeric vector indicating the ages of the nodes.
}
\description{
This function computes the crown ages of nodes on a dated phylogenetic tree.
}
\examples{
require(ape)
tree <- read.tree(text ="((t1:1,t2:1)N2:1,(t3:1,t4:1)N3:1)N1;")
plot(tree)
nodelabels()
axisPhylo()
getNage(tree, 5)
}
\seealso{
\code{\link[ape]{node.depth.edgelength}}
}
\keyword{bioregion}
