% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_mix.R
\name{gower_distance}
\alias{gower_distance}
\title{Distances for mixed variables data set}
\usage{
gower_distance(
  data,
  method = "gower",
  idnum = NULL,
  idbin = NULL,
  idcat = NULL
)
}
\arguments{
\item{data}{A data frame or matrix object.}

\item{method}{A method to calculate the mixed variables distance}

\item{idnum}{A vector of column index of the numerical variables.}

\item{idbin}{A vector of column index of the binary variables.}

\item{idcat}{A vector of column index of the categorical variables.}
}
\description{
Distances for mixed variables data set
}
