% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hex_pie.R
\name{hex.pie}
\alias{hex.pie}
\title{Plot hexagonal pies using an omega matrix to dipict different hierarchical regions.}
\usage{
hex.pie(poly.df)
}
\arguments{
\item{poly.df}{an object of SpatialPolygonsDataframe with an omega matrix}
}
\value{
A hexagonal pies plot on a map in geographic space
}
\description{
Plot hexagonal pies using an omega matrix to dipict different hierarchical regions.
}
\examples{
require(sp)
data(africa)
wm<-africa$polys
omeg<-africa$omega
hx <- spsample(wm,type="hexagonal",n=nrow(omeg))
HexPols <- HexPoints2SpatialPolygons(hx)
nex.hex <- as(HexPols,"SpatialPolygonsDataFrame")
nex.hex@data<-data.frame(omeg[1:length(nex.hex),])
hex.pie(nex.hex)

}
\author{
Piyal Karunarathne \email{piyalkarumail@yahoo.com}
}
\keyword{and}
\keyword{mapping}
\keyword{visualization}
