% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapTraitVals.R
\name{mapTraits}
\alias{mapTraits}
\title{Map species trait values in geographic space}
\usage{
mapTraits(x, trait, FUN = sum, ...)
}
\arguments{
\item{x}{A community data object - a vector (with names matching trait data)
or a data.frame or matrix (with column names matching names in trait data)}

\item{trait}{A data.frame of species traits with a column of species names
matching species names in the community data, and another column with the trait values.}

\item{FUN}{The function used to aggregate species trait values in geographic space.
By default, if \code{FUN = sum}, the sum of all species traits per area or grid cell
is calculated.}

\item{\dots}{Further arguments passed to or from other methods.}
}
\value{
A data frame of species traits by site.
}
\description{
\code{mapTraits} add species trait values to species distribution in geographic space.
}
\examples{
require(data.table)
fdir <- system.file("NGAplants", package="phyloregion")
files <- file.path(fdir, dir(fdir))
dat <- data.frame(raster2comm(files))
dd <- data.frame(table(dat$species))
names(dd) <- c("species", "range")
trait_range <- dd$range
names(trait_range) <- dd$species
gg <- mapTraits(dat, trait = trait_range, FUN = sd)
}
\author{
Barnabas H. Daru \email{darunabas@gmail.com}
}
\keyword{phyloregion}
