% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo_beta.R
\name{match.phylo.comm}
\alias{match.phylo.comm}
\alias{read.community}
\alias{as.community}
\alias{as.community.matrix}
\alias{print.community}
\title{Match taxa and in phylogeny and community matrix}
\usage{
match.phylo.comm(phy, comm, add.genus = TRUE, trace = 1)
}
\arguments{
\item{phy}{A phylogeny}

\item{comm}{A (sparse) community data matrix}

\item{add.genus}{Add species on a genus level if possible}

\item{trace}{Print comments}
}
\description{
match.phylo.comm compares taxa (species, labels, tips) present in a phylogeny
with a community matrix. Pruning, sorting and trying to add missing species
on genus level if possible to match in subsequent analysis.
}
\details{
Based on the function of the same name in picante but allows sparse matrices
and with taxa addition.
}
\keyword{cluster}
