% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matpsim.R
\name{matpsim}
\alias{matpsim}
\title{A function to calculate phylogenetic beta diversity.}
\usage{
matpsim(phyl, com, clust = 2)
}
\arguments{
\item{phyl}{Phylogenetic tree}

\item{com}{Distribution data, can be point records or range maps, represented as species × site matrix}

\item{clust}{Number of computing cores to use. Defaults to 2}
}
\description{
Using Simpson's phylogenetic beta diversity
}
\keyword{cluster}
