% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodepie.R
\name{nodepie}
\alias{nodepie}
\title{Label phylogenetic nodes using pie}
\usage{
nodepie(pie, radius = 2, pie_control = list(), legend = FALSE, ...)
}
\arguments{
\item{pie}{Estimates from ancestral character reconstruction}

\item{radius}{Radius of the pie}

\item{pie_control}{The list of control parameters to be passed into
the add.pie function.}

\item{legend}{Logical, whether to add a legend or not.}

\item{\dots}{Further arguments passed to or from other methods.}
}
\value{
Returns no value, just add color pies on phylogenetic nodes!
}
\description{
Label phylogenetic nodes using pie
}
\examples{
\donttest{
old.par <- par(no.readonly = TRUE)
require(ape)
data(africa)
m <- ace_multi(africa$theta, africa$phylo)

par(mai=rep(0,4))
par(mfrow=c(2,1))
plot(m$tree, type = "fan", show.tip.label=FALSE,
     open.angle = 180, edge.width=0.5)

nodepie(m$ancestral_states, radius = 2)

plot_structure(africa$omega, shp = africa$polys, legend = TRUE)
par(old.par)
}

}
