% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd.R
\name{PD}
\alias{PD}
\title{Phylogenetic diversity and faster}
\usage{
PD(x, phy)
}
\arguments{
\item{x}{a community matrix, i.e. an object of class matrix or Matrix.}

\item{phy}{a phylogenetic tree (object of class phylo).}
}
\value{
a vector with the PD for all samples.
}
\description{
\code{PD} calculates Faith's (1992) phylogenetic diversity and much faster
using sparse matrix.
}
\examples{
library(ape)
tree <- read.tree(text ="((t1:1,t2:1)N2:1,(t3:1,t4:1)N3:1)N1;")
com <- matrix(c(1,0,1,1,0,0,
                1,0,0,1,1,0,
                1,1,1,1,1,1,
                0,0,1,1,0,1), 6, 4,
              dimnames=list(paste0("g",1:6), tree$tip.label))
PD(com, tree)
}
\references{
Faith, D.P. (1992) Conservation evaluation and phylogenetic diversity.
\emph{Biological Conservation} \strong{61}: 1–10.
}
\seealso{
read.community read.tree phylobeta_core
}
\keyword{cluster}
