% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo_beta.R
\name{phylo_community}
\alias{phylo_community}
\alias{read.community}
\alias{as.community}
\alias{as.community.matrix}
\alias{print.community}
\alias{pd}
\title{Phylogenetic community objects}
\usage{
phylo_community(x, phy)

pd(x, tree = NULL)
}
\arguments{
\item{tree}{a phylogenetic tree (object of class phylo)}

\item{splist}{Species list}

\item{nodes}{node list}

\item{output.splist}{return species list}
}
\description{
read.community reads in file containing occurence data and returns a sparse
matrix.
}
\details{
If species is not in the tip label species is added to the genus or family
level when possible
}
\examples{
example(as.community)
pc <- phylo_community(com, tree)
pd(pc)

}
\keyword{cluster}
