% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhyloMaker.R
\name{phylo_maker}
\alias{phylo_maker}
\title{Create a subtree with largest overlap from a species list.}
\usage{
phylo_maker(splist, tree, nodes, output.splist = TRUE)
}
\arguments{
\item{splist}{Species list}

\item{tree}{a phylogenetic tree (object of class phylo)}

\item{nodes}{node list}

\item{output.splist}{return species list}
}
\description{
Create a subtree with largest overlap from a species list.
}
\details{
If species is not in the tip label species is added to the genus or family
level when possible
}
\keyword{cluster}
