% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo_beta.R
\name{phylobeta.core}
\alias{phylobeta.core}
\title{Phylogenetic beta diversity}
\usage{
phylobeta.core(x, phy, index.family = "sorensen", index = "Simpson")
}
\arguments{
\item{x}{an object of class Matrix or matrix}

\item{phy}{a phylogenetic tree (object of class phylo)}
}
\description{
\code{beta.core} and \code{phylobeta.core} computes efficiently for
large community matrices and trees the necessary quantities used by the
betapart package to compute pairwise and multiple-site phylogenetic
dissimilarities.
}
\examples{
library(ape)
tree <- read.tree(text ="((t1:1,t2:1)N2:1,(t3:1,t4:1)N3:1)N1;")
com <- matrix(c(1,0,1,1,0,0,
                1,0,0,1,1,0,
                1,1,1,1,1,1,
                0,0,1,1,0,1), 6, 4,
              dimnames=list(paste0("g",1:6), tree$tip.label))
pbc <- phylobeta.core(com, tree)

}
\seealso{
read.community PD
}
\keyword{phyloregion}
