% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo_beta.R
\name{phylobeta_core}
\alias{phylobeta_core}
\alias{phylo_betapart_core}
\alias{beta_sparse_core}
\title{Phylogenetic beta diversity}
\usage{
phylobeta_core(x, phy)

beta_sparse_core(x)
}
\arguments{
\item{x}{an object of class Matrix or matrix}

\item{phy}{a phylogenetic tree (object of class phylo)}
}
\description{
\code{beta_sparse_core} and \code{phylobeta_core} computes efficiently for
large community matrices and trees the nesseary quantitities used by the
betapart package to compute pairwise and multiple-site phylogenetic
dissimilarities.
}
\examples{
library(ape)
tree <- read.tree(text ="((t1:1,t2:1)N2:1,(t3:1,t4:1)N3:1)N1;")
com <- matrix(c(1,0,1,1,0,0,
                1,0,0,1,1,0,
                1,1,1,1,1,1,
                0,0,1,1,0,1), 6, 4,
              dimnames=list(paste0("g",1:6), tree$tip.label))
pbc <- phylobeta_core(com, tree)
library(betapart)
phylo.beta.multi(pbc)
phylo.beta.pair(pbc)

}
\seealso{
read.community pd
}
\keyword{cluster}
