% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo_builder.R
\name{phylobuilder}
\alias{phylobuilder}
\title{Create a subtree with largest overlap from a species list.}
\usage{
phylobuilder(species, tree)
}
\arguments{
\item{species}{A vector or matrix containing a species list}

\item{tree}{a phylogenetic tree (object of class phylo)}
}
\description{
phylobuilder creates a subtree with largest overlap from a species list.
If species in the species list are not already in the tip label, species will
be added at the most recent common ancestor at the genus or family level when
possible.
}
\examples{
data("africa")
phylobuilder(species = colnames(africa$comm), tree = africa$phylo)

library(ape)
cats_and_dogs <- read.tree(text="(((((Panthera_leo,Panthera_pardus),
  Panthera_onca),(Panthera_uncia,(Panthera_tigris_altaica,
  Panthera_tigris_amoyensis)))Panthera)Felidae,(((((((Canis_lupus,
  Canis_lupus_familiaris),Canis_latrans),Canis_anthus),Canis_aureus),
  Lycaon_pictus),(Canis_adustus,Canis_mesomelas))Canis,(Vulpes_vulpes,
  Vulpes_corsac)Vulpes)Canidae)Carnivora;")
plot(cats_and_dogs, show.node.label=TRUE)



}
\seealso{
\code{\link[phangorn]{add.tips}}, \code{\link[ape]{label2table}},
\code{\link[ape]{stripLabel}}
}
\keyword{bioregion}
