% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloregion-package.R
\docType{package}
\name{phyloregion-package}
\alias{phyloregion-package}
\alias{phyloregion}
\title{Biogeographic regionalization and spatial conservation}
\description{
Biogeographic regionalization and spatial conservation
}
\author{
Barnabas Daru, Piyal Karunarathne, Klaus Schliep
}
\keyword{package}
