% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/see_sparse.R
\name{plot.sparse}
\alias{plot.sparse}
\title{Create illustrative sparse matrix}
\usage{
\method{plot}{sparse}(x, col = c("red", "yellow"), lwd = 1, ...)
}
\arguments{
\item{x}{A matrix}

\item{col}{A vector of colors to represent presence or absence of a species}

\item{lwd}{Line width}

\item{\dots}{Further arguments passed to or from other methods.}
}
\value{
Returns no value, just plot sparse matrix
}
\description{
This function visualizes a sparse matrix using vertical bands corresponding
to presence or absence of a species in an area.
}
