% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_NMDS.R
\name{plot_NMDS}
\alias{plot_NMDS}
\title{Plot phyloregions in NMDS ordinatoion space}
\usage{
plot_NMDS(x, k = 10, method = "average", ...)
}
\arguments{
\item{x}{a pairwise matrix of beta diversity (phylo or non-phylo).}

\item{k}{the optimal number of clusters derived from the elbow method
(as in the \code{GMD} package).}

\item{method}{the agglomeration method to be used. This should
be (an unambiguous abbreviation of) one of \dQuote{ward.D}, \dQuote{ward.D2},
\dQuote{single}, \dQuote{complete}, \dQuote{average} (= UPGMA), \dQuote{mcquitty} (= WPGMA),
\dQuote{median} (= WPGMC) or \dQuote{centroid} (= UPGMC).}

\item{...}{graphical parameters to plot}
}
\value{
Returns no value, just plot the bioregions
}
\description{
This function plots clusters of cells in NMDS space.
}
\examples{
data(africa)
tree <- africa$phylo
x <- sampl2sparse(africa$comm)

subphy <- match_phylo_comm(tree, x)$phy
submat <- match_phylo_comm(tree, x)$com

pbc <- phylobeta(submat, subphy)
plot_NMDS(pbc)
}
\author{
Barnabas H. Daru
}
\keyword{and}
\keyword{mapping}
\keyword{visualization}
