% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBioregion.R
\name{plot_phyloregion}
\alias{plot_phyloregion}
\title{Map phyloregions and bioregions}
\usage{
plot_phyloregion(
  dat,
  shp,
  method = "average",
  k = 15,
  border = NA,
  cex = 1,
  ...
)
}
\arguments{
\item{dat}{a pairwise matrix of beta diversity (phylo or non-phylo).}

\item{shp}{a polygon shapefile of grid cells.}

\item{method}{the agglomeration method to be used. This should
be (an unambiguous abbreviation of) one of \dQuote{ward.D}, \dQuote{ward.D2},
\dQuote{single}, \dQuote{complete}, \dQuote{average} (= UPGMA), \dQuote{mcquitty} (= WPGMA),
\dQuote{median} (= WPGMC) or \dQuote{centroid} (= UPGMC).}

\item{k}{the optimal number of clusters derived from the elbow method
(as in the \code{GMD} package).}

\item{border}{color of the border.}

\item{cex}{character (or symbol) expansion: a numerical vector.}

\item{...}{graphical parameters to plot}
}
\value{
Returns no value, just plot the bioregions
}
\description{
This function plots clusters of cells i.e \sQuote{phyloregions} or \sQuote{bioregions} in
geographic space based on turnover of species or phylogenetic lineages.
}
\examples{
data(africa)
tree <- africa$phylo
x <- sampl2sparse(africa$comm)

subphy <- match_phylo_comm(tree, x)$phy
submat <- match_phylo_comm(tree, x)$com

pbc <- phylobeta(submat, subphy)
plot_phyloregion(pbc, shp=africa$polys)
}
\references{
Daru, B.H., Van der Bank, M., Maurin, O., Yessoufou, K., Schaefer, H., Slingsby, J.A. & Davies, T.J. (2016) A novel phylogenetic
regionalization of the phytogeographic zones of southern Africa reveals their hidden evolutionary affinities.
\emph{Journal of Biogeography} \strong{43}: 155-166.


Daru, B.H., Elliott, T.L., Park, D.S. & Davies, T.J. (2017) Understanding the processes underpinning patterns of phylogenetic
regionalization. \emph{Trends in Ecology and Evolution} \strong{32}: 845-860.
}
\author{
Barnabas H. Daru
}
\keyword{and}
\keyword{mapping}
\keyword{visualization}
