% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pie.R
\name{plot_pie}
\alias{plot_pie}
\title{Visualize biogeographic patterns using pie charts}
\usage{
plot_pie(x = NULL, shp.grids, K = 5, pie_control = list(), ...)
}
\arguments{
\item{x}{a matrix of phyloregion of probabilities of each species}

\item{shp.grids}{if specified, the polygon shapefile of grid cells
with a column labeled \dQuote{grids}.}

\item{K}{Number of distinctive colors for the pies corresponding to the
number of clusters or regions.}

\item{pie_control}{The list of control parameters to be passed into
the add.pie function.}

\item{\dots}{Further arguments passed to or from other methods.}
}
\value{
Returns no value, just map color pies in geographic space!
}
\description{
Visualize biogeographic patterns using pie charts
}
