% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pie.R
\name{plot_structure}
\alias{plot_structure}
\title{Visualize biogeographic patterns using pie charts}
\usage{
plot_structure(
  x = NULL,
  shp,
  r = 1,
  pie_control = list(),
  legend = FALSE,
  legend_pie = FALSE,
  ...
)
}
\arguments{
\item{x}{a matrix of phyloregion of probabilities of each species}

\item{shp}{if specified, the polygon shapefile of grid cells
with a column labeled \dQuote{grids}.}

\item{r}{Radius of the pie legend to be displayed}

\item{pie_control}{The list of control parameters to be passed into
the add.pie function.}

\item{legend}{Logical, whether to plot a legend or not.}

\item{legend_pie}{Legend for the pie plots.}

\item{\dots}{Further arguments passed to or from other methods.}
}
\value{
Returns no value, just map color pies in geographic space!
}
\description{
Visualize biogeographic patterns using pie charts
}
\examples{
data(africa)
plot_structure(africa$omega, shp = africa$polys)
}
