% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.swatch.R
\name{plot_swatch}
\alias{plot_swatch}
\title{Plot shapefile polygons based on slot values}
\usage{
plot_swatch(
  x,
  values,
  k = 10,
  swatch = "Blue-Red 3",
  lab = "",
  leg = 5,
  lwd = 15,
  pos = "bottomright",
  legend = TRUE,
  ...
)
}
\arguments{
\item{x}{A data frame or object of the class SpatialPolygonsDataFrame}

\item{values}{Variable in the SpatialPolygonsDataFrame for which to discretize
the values of the quantity.}

\item{k}{Numeric, the desired number of bins to discretize.}

\item{swatch}{name of the palette to generate colors from. The name is matched
to the list of available color palettes from the code{hcl.colors} function in
the code{grDevices} package.}

\item{leg}{Numeric, length of the legend}

\item{lwd}{numeric, line width of the legend.}

\item{pos}{location to position the legend such as \dQuote{bottomright},
\dQuote{bottomleft}, \dQuote{topleft}, and \dQuote{topright}.}

\item{legend}{logical indicating whether to add a legend to the map.}

\item{\dots}{Further arguments passed to or from other methods.}

\item{title}{a character string indicating the caption to be placed on top
of the legend.}
}
\value{
Returns no value, just map swatch of colors in geographic space!
}
\description{
\code{plot.swatch} maps discretized values of a quantity based on their quantiles.
}
\examples{
require(raster)
s <- readRDS(system.file("nigeria/SR_Naija.rds", package= "bioregion"))
plot_swatch(s, values = s$SR, k=20)
}
\author{
Barnabas H. Daru \email{darunabas@gmail.com}
}
\keyword{and}
\keyword{mapping}
\keyword{visualization}
