% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points2comm.R
\name{points2comm}
\alias{points2comm}
\title{Convert point occurrences to community data}
\usage{
points2comm(
  dat,
  mask = NULL,
  res = 1,
  lon = "decimallongitude",
  lat = "decimallatitude",
  species = "species",
  shp.grids = NULL,
  index = "taxon_richness",
  ...
)
}
\arguments{
\item{dat}{A point occurrence data frame with at least three columns:
\itemize{
  \item Column 1: \code{species} (listing the taxon names)
  \item Column 2: \code{decimallongitude} (corresponding to decimal longitude)
  \item Column 3: \code{decimallatitude} (corresponding to decimal latitude)
}}

\item{mask}{a polygon shapefile covering the boundary of the survey region.}

\item{res}{the grain size of the grid cells in decimal degrees (default).}

\item{lon}{character with the column name of the longitude.}

\item{lat}{character with the column name of the lattitude.}

\item{species}{a character string. The column with the species or taxon name.
Default = \dQuote{species}.}

\item{shp.grids}{if specified, the polygon shapefile of grid cells
with a column labeled \dQuote{grids}.}

\item{index}{Diversity index, one of \dQuote{taxon_richness} (default) or
\dQuote{abundance}.}

\item{\dots}{Further arguments passed to or from other methods.}
}
\value{
\itemize{
  \item comm_dat: community data frame
  \item poly_shp: shapefile of grid cells with the values per cell.
}
}
\description{
\code{points2comm} converts point occurrence records (e.g. from GBIF,
iDigBio) to a community data sample
}
\examples{
require(raster)
s <- readRDS(system.file("nigeria/nigeria.rds", package="phyloregion"))

set.seed(1)
m <- data.frame(sp::spsample(s, 10000, type="nonaligned"))
names(m) <- c("lon", "lat")
species <- paste0("sp", sample(1:1000))
m$taxon <- sample(species, size = nrow(m), replace = TRUE)
pts <- points2comm(dat = m, mask = s, res = 0.5, lon="lon", lat="lat", species="taxon")

plot_swatch(pts$poly_shp, values = pts$poly_shp$richness, k=10)
}
\author{
Barnabas H. Daru \email{darunabas@gmail.com}
}
\keyword{phyloregion}
