% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress.R
\name{progress}
\alias{progress}
\title{Progress Bar}
\usage{
progress(x, FUN, ...)
}
\arguments{
\item{x}{a vector (list) or an expression object.}

\item{FUN}{the function to be applied to each element of x.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
For lapply, sapply(simplify = FALSE) and replicate(simplify = FALSE), a list.
}
\description{
This function shows the progress bar of analysis.
}
\examples{
\donttest{
m <- sapply(1:50000, function(x) list(rnorm(1000)))
n <- progress(m, sd)
}
}
