% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_species.R
\name{random_species}
\alias{random_species}
\title{Generate random species distributions in space}
\usage{
random_species(n, species, pol, ...)
}
\arguments{
\item{n}{vector of one or more elements to choose from, or a positive
integer.}

\item{species}{the desired number of species.}

\item{pol}{the vector polygon of the study area for determining the
species distributions}

\item{\dots}{Further arguments passed to or from other methods.}
}
\value{
A vector polygon of species' extent of occurrence ranges.
}
\description{
This function generates random species distributions in geographic space as
extent of occurrence range polygons based on convex hulls of random points.
}
\author{
Barnabas H. Daru \email{darunabas@gmail.com}
}
\keyword{phyloregion}
