% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rast_quantile.R
\name{rast_quantile}
\alias{rast_quantile}
\title{Standardizes raster values for mapping}
\usage{
rast_quantile(ras, ref)
}
\arguments{
\item{ras}{an input raster layer.}

\item{ref}{a raster layer for reference.}
}
\value{
A raster layer that has been standardized and ready for mapping
}
\description{
This function standardizes values of a raster layer for mapping.
}
