% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster2comm.R
\name{raster2comm}
\alias{raster2comm}
\title{Convert raster to community matrix}
\usage{
raster2comm(files)
}
\arguments{
\item{files}{list of raster layer objects with the same spatial 
extent and resolution.}
}
\description{
\code{raster2comm} converts a set of raster files from a species distribution
modeling into a community matrix.
}
\examples{
fdir <- system.file("NGAplants", package="phyloregion")
files <- file.path(fdir, dir(fdir))
r <- raster2comm(files)

}
\keyword{bioregion}
