% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community.R
\name{read.community}
\alias{read.community}
\title{Read in sparse community matrices}
\usage{
read.community(file, ...)
}
\arguments{
\item{file}{A file name.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
\code{read.community} reads in file containing occurence data and returns a
sparse matrix.
}
\examples{
data(africa)
}
\keyword{cluster}
