% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample2sparsematrix.R
\name{sampl2sparse}
\alias{sampl2sparse}
\title{Convert community data to sparse matrix.}
\usage{
sampl2sparse(dat, grids = "grids", species = "species", method = "phylo")
}
\arguments{
\item{dat}{A community data frame with two columns, grids and species
(or taxa)}

\item{method}{one of "phylo" (the default) corresponding to phylogenetic
beta diversity, or "nonphylo" for beta diversity.}
}
\value{
A compressed sparse community matrix of sites by species
}
\description{
This function converts community data frame to compressed sparse matrix.
It is especially developed for handling big data.
}
\examples{
require(data.table)
fdir <- system.file("NGAplants", package="phyloregion")
files <- file.path(fdir, dir(fdir))
x <- data.frame(raster2comm(files))
comm <- sampl2sparse(x)
}
\keyword{bioregion}
