% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdm_terra.R
\name{sdmterra}
\alias{sdmterra}
\title{Species distribution models for a range of algorithms}
\usage{
sdmterra(
  x,
  predictors = NULL,
  blank = NULL,
  algorithm = "GLM",
  size = 50,
  width = 1e+05
)
}
\arguments{
\item{x}{A dataframe containing the species occurrences
and geographic coordinates. Column 1 labeled as "species", column 2 "lon",
column 3 "lat".}

\item{predictors}{RasterStack of environmental descriptors on which
the models will be projected}

\item{blank}{A blank raster upon which the prediction layer is aggregated to.}

\item{algorithm}{Character. The choice of algorithm to run the species
distribution model. Available algorithms include:
\itemize{
\item \dQuote{RF}: Random forest.
\item \dQuote{GLM}: Generalized linear model.
\item \dQuote{MAXENT}: Maximum entropy.
\item \dQuote{GBM}: Generalized boosted regressions model.}}

\item{size}{Minimum number of points required to successfully run
a species distribution model}

\item{width}{Width of buffer in meter if x is in longitude/latitude CRS.}
}
\value{
A list with the following objects:
\itemize{
  \item \code{ensemble_raster} The ensembled raster that predicts
  the potential species distribution.
  \item \code{ensemble_AUC} The median AUCs of models.
  \item \code{data} The dataframe that was used to implement the model.
  \item \code{indiv_models} Raster layers for the separate models that
  predict the potential species distribution.
  \item \code{single_AUCs} The AUCs for the seperate models.
}
}
\description{
This function computes species distribution models using
four modelling algorithms: generalized linear models,
generalized boosted models, random forests, and maximum entropy (only if
\code{rJava} is available). Note: this is an experimental function, and
may change in the future.
}
\references{
Zurell, D., Franklin, J., König, C., Bouchet, P.J., Dormann, C.F., Elith, J.,
Fandos, G., Feng, X., Guillera‐Arroita, G., Guisan, A., Lahoz‐Monfort, J.J.,
Leitão, P.J., Park, D.S., Peterson, A.T., Rapacciuolo, G., Schmatz, D.R.,
Schröder, B., Serra‐Diaz, J.M., Thuiller, W., Yates, K.L., Zimmermann, N.E.
and Merow, C. (2020), A standard protocol for reporting species distribution
models. \emph{Ecography}, \strong{43}: 1261-1277.
}
