% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectbylocation.R
\name{selectbylocation}
\alias{selectbylocation}
\title{Select features within polygon from another layer}
\usage{
selectbylocation(x, y)
}
\arguments{
\item{x}{source layer of the class SpatialPolygonsDataFrame or SpatialPointsDataFrame}

\item{y}{Target layer or mask extent to subset from.}
}
\value{
A spatial polygons or spatial points object pruned to the extent 
of the target layer.
}
\description{
The \code{selectbylocation} function selects features based on 
their location relative to features in another layer.
}
\examples{
require(raster)
require(maptools)
data("wrld_simpl")
set.seed(1)
m <- data.frame(lon = runif(2000, -180, 180), 
                lat = runif(2000, -90, 90), sites=seq(2000))
coordinates(m) <- ~lon+lat
z <- selectbylocation(m, wrld_simpl)
plot(wrld_simpl)
points(m, col="blue", pch="+")
points(z, col="red", pch="+")
}
\keyword{bioregion}
