% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparsematrix2sampl.R
\name{sparse2sampl}
\alias{sparse2sampl}
\title{Deconstruct sparse matrix to community data}
\usage{
sparse2sampl(x)
}
\arguments{
\item{x}{A compressed sparse matrix of class \dQuote{dgCMatrix}, or \dQuote{dgRMatrix}.}
}
\value{
a community data frame of sites and species pairs
}
\description{
This function converts a compressed sparse matrix to a base R
community data frame of (row, column) pairs.
}
\examples{
data(africa)
tree <- africa$phylo
M <- sampl2sparse(africa$comm, method = "nonphylo")
N <- sparse2sampl(M)
}
\keyword{phyloregion}
