% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait_matrix.R
\name{trait_matrix}
\alias{trait_matrix}
\title{Phyloregions for functional traits}
\usage{
trait_matrix(
  x,
  trait,
  num = NULL,
  bin = NULL,
  cat = NULL,
  cut = NULL,
  phy = NULL,
  method = "trait"
)
}
\arguments{
\item{x}{A community data in long format with one column representing
sites labeled \dQuote{grids} and another column representing species
labeled \dQuote{species}.}

\item{trait}{A data frame or matrix object with the first column
labeled \dQuote{species} containing the texonomic groups to be
evaluated whereas the remaining columns have the various functional
traits. The variables must be a mix of numeric and categorical values.}

\item{num}{A vector of column index of the numerical trait variable.}

\item{bin}{A vector of column index of the binary trait variable.}

\item{cat}{A vector of column index of the categorical trait variable.}

\item{cut}{The slice time}

\item{phy}{is a dated phylogenetic tree with branch lengths stored
as a phylo object (as in the ape package).}

\item{method}{Whether to compute phylogenetic
(method = \dQuote{phylo}), functional (method = \dQuote{trait})
or on single functional categorical variables (method =
\dQuote{single_cat})}
}
\value{
Function returns a community data frame that captures the
count of each species based on its cluster membership.
}
\description{
Generates a sparse community matrix as input for clustering regions
based on the similairity of functional traits across species.
}
\keyword{bioregion}
