% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait_network.R
\name{trait_network}
\alias{trait_network}
\title{Map network of functional traits}
\usage{
trait_network(x, d, k = 5)
}
\arguments{
\item{x}{A data frame, matrix or edgelist}

\item{d}{A data frame with the functional traits}

\item{k}{The number to bin discrete traits.}
}
\value{
{A graph object}
}
\description{
This function presents a network approach to functional beta diversity
}
\examples{
com <- data.frame(grids = c("v1", "v1", "v2", "v2", "v3", "v3", "v3"),
                 species = c("s1", "s3", "s2", "s3", "s1", "s2", "s3"))

trait <- data.frame(species = c("s1", "s2", "s3"),
                   habit = c("tree", "herb", "liana"),
                   IUCN = c("LC", "VU", "CR"),
                   height = c(7.9, 0.5, 3))

g <- trait_network(x=com, d=trait)
}
\keyword{phyloregion}
