% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeSampler.R
\name{tree.sampler}
\alias{tree.sampler}
\title{Subset trees from posterior distribution of trees.}
\usage{
tree.sampler(wd, n = 100, ...)
}
\arguments{
\item{wd}{A path to the working directory with the distributions of 
multiple phylogenetic trees.}

\item{n}{The desired number of subsets of trees. This defaults to 100.}

\item{\dots}{arguments passed among methods.}
}
\value{
An object of class "multiPhylo" with the subset of trees.
}
\description{
This function randomly samples a subset of trees from a posterior 
distribution of trees derived from multiple runs of MrBayes.
}
\author{
Dominic Bennett & Harith Farooq \email{harithmorgadinho@gmail.com}
}
