% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeightedEndemism.R
\name{weighted.endemism}
\alias{weighted.endemism}
\title{Measure the distribution of narrow-ranged or endemic species.}
\usage{
weighted.endemism(x)
}
\arguments{
\item{x}{A community matrix or data frame.}
}
\value{
A data frame of species traits by site.
}
\description{
\code{weighted.endemism} is species richness inversely weighted
by species ranges.
}
\examples{
require(raster)
data(africa)

Endm <- weighted.endemism(africa$comm)
m <- merge(africa$polys, Endm, by="grids")
m <- m[!is.na(m@data$WE),]

plot.swatch(m, values = m$WE, k=20)

}
\references{
Crisp, M.D., Laffan, S., Linder, H.P. & Monro, A. (2001) Endemism in the Australian flora.
\emph{Journal of Biogeography} \strong{28}: 183–198.

Laffan, S.W., & Crisp, M.D. (2003) Assessing endemism at multiple spatial scales,
with an example from the Australian vascular flora. \emph{Journal of Biogeography} \strong{30}: 511–520.
}
\keyword{bioregion}
