% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Choropleth.R
\name{choropleth}
\alias{choropleth}
\title{Bin shapefile polygons based on slot values}
\usage{
choropleth(x, values, k = 10, ...)
}
\arguments{
\item{x}{A data frame or object of the class SpatialPolygonsDataFrame}

\item{values}{Variable in the SpatialPolygonsDataFrame for which to
discretize the values of the quantity.}

\item{k}{Numeric, the desired number of bins to discretize.}

\item{\dots}{Further arguments passed to or from other methods.}
}
\value{
{returns a SpatialPolygonsDataFrame with a column of the discretized values}
}
\description{
\code{choropleth} discretizes the values of a quantity based on their
quantiles.
}
\examples{
require(raster)
s <- readRDS(system.file("nigeria/SR_Naija.rds", package = "phyloregion"))
k = 10
COLOUR <- hcl.colors(k, "RdYlBu")
y <- choropleth(s, values = s$SR, k)

## To plot and color according to some metric:
plot(y, col = COLOUR[y$values], border = NA)
}
\author{
Barnabas H. Daru \email{darunabas@gmail.com}
}
\keyword{bioregion}
