% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo_beta.R
\name{match_phylo_comm}
\alias{match_phylo_comm}
\title{Match taxa and in phylogeny and community matrix}
\usage{
match_phylo_comm(phy, comm)
}
\arguments{
\item{phy}{A phylogeny}

\item{comm}{A (sparse) community data matrix}

\item{trace}{Print comments}
}
\description{
match_phylo_comm compares taxa (species, labels, tips) present in a phylogeny
with a community matrix. Pruning, sorting and trying to add missing species
on genus level if possible to match in subsequent analysis.
}
\details{
Based on the function of the same name in picante but allows sparse matrices
and with taxa addition.
}
\examples{
example(phylo_builder)
pc <- match_phylo_comm(tree, comm)
}
\keyword{cluster}
