% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_matrix.R
\name{mean_dist}
\alias{mean_dist}
\title{Mean distance matrix from a set of distance matrices}
\usage{
mean_dist(files, trace = 1, ...)
}
\arguments{
\item{files}{list of pairwise distance matrices stored as CSVs or .rds
with the same dimensions.}

\item{trace}{Trace the function; trace = 2 or higher will be more voluminous.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
average distance matrix
}
\description{
This function generates the mean pairwise distance matrix from a set
many pairwise distance matrices. Note: all matrices should be of the same dimension.
}
