% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodepie.R
\name{nodepie}
\alias{nodepie}
\title{Label phylogenetic nodes using pie}
\usage{
nodepie(
  pie,
  radius = 2,
  pie_control = list(),
  legend = FALSE,
  col = hcl.colors(5),
  ...
)
}
\arguments{
\item{pie}{Estimates from ancestral character reconstruction}

\item{radius}{Radius of the pie}

\item{pie_control}{The list of control parameters to be passed into
the add.pie function.}

\item{legend}{Logical, whether to add a legend or not.}

\item{col}{List of colors for the pies.}

\item{\dots}{Further arguments passed to or from other methods.}
}
\value{
Returns no value, just add color pies on phylogenetic nodes!
}
\description{
Label phylogenetic nodes using pie
}
