% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo_builder.R
\name{phylo_builder}
\alias{phylo_builder}
\title{Create a subtree with largest overlap from a species list.}
\usage{
phylo_builder(species, tree, output.species_list = TRUE, level = c("family",
  "genus", "species"))
}
\arguments{
\item{tree}{a phylogenetic tree (object of class phylo)}

\item{output.species_list}{return species list}

\item{species_list}{Species list}
}
\description{
phylobuilder creates a subtree with largest overlap from a species list.
If species in the species list are not already in the tip label, species will
added at the most recent commaon ancestor at the genus or family level when
possible.
}
\keyword{cluster}
