% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo_beta.R
\name{phylo_community}
\alias{phylo_community}
\alias{pd}
\alias{phylo_betapart_core}
\alias{pd}
\alias{phylo_betapart_core}
\title{Phylogenetic community objects}
\usage{
phylo_community(x, phy)

pd(x, tree = NULL)

phylo_betapart_core(x)
}
\arguments{
\item{x}{an object of class Matrix or phylo_community}

\item{phy}{a phylogenetic tree (object of class phylo)}
}
\description{
read.community reads in file containing occurence data and returns a sparse
matrix.
}
\details{
If species is not in the tip label species is added to the genus or family
level when possible
}
\examples{
tree <- read.tree(text ="((t1:1,t2:1)N2:1,(t3:1,t4:1)N3:1)N1;")
com <- matrix(c(1,0,1,1,0,0,
                1,0,0,1,1,0,
                1,1,1,1,1,1,
                0,0,1,1,0,1), 6, 4,
              dimnames=list(paste0("g",1:6), tree$tip.label))
pc <- phylo_community(com, tree)
pd(pc)
pbc <- phylo_betapart_core(pc)
library(betapart)
phylo.beta.multi(pbc)
phylo.beta.pair(pbc)

}
\keyword{cluster}
