% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo_beta.R
\name{phylo_community}
\alias{phylo_community}
\alias{pd}
\alias{phylo_betapart_core}
\alias{phylobeta_core}
\alias{beta_sparse_core}
\title{Phylogenetic community objects}
\usage{
phylo_community(x, phy)

pd(x, phy = NULL)

phylobeta_core(x)

beta_sparse_core(x)
}
\arguments{
\item{x}{an object of class Matrix or phylo_community}

\item{phy}{a phylogenetic tree (object of class phylo)}
}
\description{
phylo_community converts a community matrix and a tree in phylogenetic
community matrix. This objects allows for efficient computation of
phylogenetic diversity (pd) and beta diversity indices (phylo_betapart_core).
}
\examples{
library(ape)
tree <- read.tree(text ="((t1:1,t2:1)N2:1,(t3:1,t4:1)N3:1)N1;")
com <- matrix(c(1,0,1,1,0,0,
                1,0,0,1,1,0,
                1,1,1,1,1,1,
                0,0,1,1,0,1), 6, 4,
              dimnames=list(paste0("g",1:6), tree$tip.label))
pc <- phylo_community(com, tree)
pd(pc)
pbc <- phylobeta_core(pc)
library(betapart)
phylo.beta.multi(pbc)
phylo.beta.pair(pbc)

}
\seealso{
read.community
}
\keyword{cluster}
