% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloregion-package.R
\docType{package}
\name{phyloregion-package}
\alias{phyloregion-package}
\alias{phyloregion}
\title{Biogeographic regionalization and spatial conservation}
\description{
This R package is for analysis of biogeographic regionalization (the
classification of geographical areas in terms of their biotas) and spatial
conservation in the R scientific computing environment. Previously it was
only possible to perform analysis of biogeographic regionalization on small
datasets, often using tools that are difficult to replicate. With
macroecological datasets of ever increasing size and complexity,
`phyloregion` now offers the possibility of clustering large-scale species
distributions combined with phylogenetic information using sparse matrix,
determining optimal number of clusters, quantifying evolutionary
distinctiveness of bioregions, and visualizing patterns and underlying
processes. It is simple, fast, and particularly tailored for handling large
datasets.
}
\author{
Barnabas Daru, Piyal Karunarathne, Klaus Schliep
}
\keyword{package}
