% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotED.R
\name{plot_evoldistinct}
\alias{plot_evoldistinct}
\title{Map evolutionary distinctiveness of phyloregions in space}
\usage{
plot_evoldistinct(
  dat,
  method = "average",
  shp = shp,
  k = 10,
  cex = 1,
  swatch = "YlOrBr",
  pos = "bottomright",
  lab = "",
  legend = TRUE,
  border = NA,
  leg = 5,
  lwd = 15,
  ...
)
}
\arguments{
\item{dat}{a pairwise matrix of beta diversity (phylo or non-phylo).}

\item{method}{the agglomeration method to be used. This can be one
of "ward.D", "ward.D2", "single", "complete", "average" (= UPGMA),
"mcquitty" (= WPGMA), "median" (= WPGMC) or "centroid" (= UPGMC).}

\item{shp}{a polygon shapefile of grid cells.}

\item{k}{the optimal number of clusters derived from the elbow method
(as in the \code{GMD} package).}

\item{cex}{character (or symbol) expansion: a numerical vector.}

\item{swatch}{name of the palette to generate colors from. The name is matched
to the list of available color palettes from the code{hcl.colors} function in
the code{grDevices} package.}

\item{pos}{location to position the legend such as \dQuote{bottomright},
\dQuote{bottomleft}, \dQuote{topleft}, and \dQuote{topright}.}

\item{legend}{logical indicating whether to add a legend to the map.}

\item{border}{color of the border.}

\item{\dots}{arguments passed among methods.}

\item{title}{a character string indicating the caption to be placed on top
of the legend.}
}
\value{
Returns no value, just plot the evolutionary distinctiveness of bioregions
}
\description{
This function maps evolutionary distinctiveness of bioregions based on the
mean value of phylogenetic beta diversity between a focal bioregion and all
other bioregions in the study area.
}
\examples{
data(africa)
tree <- africa$phylo
x <- sampl2sparse(africa$comm)

subphy <- match_phylo_comm(tree, x)$phy
submat <- match_phylo_comm(tree, x)$com

pbc <- phylobeta(submat, subphy)
plot_evoldistinct(pbc, shp=africa$polys)
}
\references{
Daru, B.H., Van der Bank, M., Maurin, O., Yessoufou, K., Schaefer, H., Slingsby, J.A. & Davies, T.J. (2016) A novel phylogenetic
regionalization of the phytogeographic zones of southern Africa reveals their hidden evolutionary affinities.
\emph{Journal of Biogeography} \strong{43}: 155-166.

Daru, B.H., Elliott, T.L., Park, D.S. & Davies, T.J. (2017) Understanding the processes underpinning patterns of phylogenetic
regionalization. \emph{Trends in Ecology and Evolution} \strong{32}: 845-860.
}
\author{
Barnabas H. Daru
}
