% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_structure.R
\name{plot_structure}
\alias{plot_structure}
\title{Visualize biogeographic patterns using structure plots}
\usage{
plot_structure(omega, shp, by = NULL, col = hcl.colors(15), ...)
}
\arguments{
\item{omega}{a matrix of phyloregion of probabilities of each species}

\item{shp}{a polygon shapefile of grid cells with a column labeled
\dQuote{grids}.}

\item{by}{Specification of the column used for grouping.}

\item{col}{List of colors for the pies.}

\item{\dots}{Further arguments passed to or from other methods.}
}
\value{
Returns no value, just visualize structure plot!
}
\description{
Visualize biogeographic patterns using structure plots
}
