% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community.R
\name{read.community}
\alias{read.community}
\title{Read in sparse community matrices}
\usage{
read.community(file, ...)
}
\arguments{
\item{splist}{Species list}

\item{tree}{a phylogenetic tree (object of class phylo)}

\item{nodes}{node list}

\item{output.splist}{return species list}
}
\description{
read.community reads in file containing occurence data and returns a sparse
matrix.
}
\details{
If species is not in the tip label species is added to the genus or family
level when possible
}
\keyword{cluster}
