% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdm_terra.R
\name{sdmterra}
\alias{sdmterra}
\title{Species distribution models for a range of algorithms}
\usage{
sdmterra(
  x,
  predictors = NULL,
  blank = NULL,
  algorithm = "all",
  size = 50,
  width = 1e+05
)
}
\arguments{
\item{x}{A dataframe containing the species occurrences
and geographic coordinates. Column 1 labeled as "species", column 2 "lon",
column 3 "lat".}

\item{predictors}{A \code{SpatRaster} to extract values from the
locations in x on which the models will be projected}

\item{blank}{A blank raster upon which the prediction layer is aggregated to.
If not specified, the first \code{SpatRaster} layer from \code{predictors}
above is used as blank.}

\item{algorithm}{Character. The choice of algorithm to run the species
distribution model. Available algorithms include:
\itemize{
\item \dQuote{all}: Calls all available algorithms: RF, GLM, MAXENT, GBM,
and CTA.
\item \dQuote{RF}: Calls only Random forest.
\item \dQuote{GLM}: Calls only Generalized linear model.
\item \dQuote{MAXENT}: Calls only Maximum entropy.
\item \dQuote{GBM}: Calls only Generalized boosted regressions model.
\item \dQuote{CTA}: Calls only Classification tree analysis.}}

\item{size}{Minimum number of points required to successfully run
a species distribution model especially for species with few occurrences.}

\item{width}{Width of buffer in meter if x is in longitude/latitude CRS.}
}
\value{
A list with the following objects:
\itemize{
  \item \code{ensemble_raster} The ensembled raster that predicts
  the potential species distribution.
  \item \code{data} The dataframe of occurrences used to implement the model.
  \item \code{indiv_models} Raster layers for the separate models that
  predict the potential species distribution.
}
}
\description{
This function computes species distribution models using
five modelling algorithms: generalized linear models,
generalized boosted models, random forests, maximum entropy (only if
\code{rJava} is available), and classification tree analysis.
Note: this is an experimental function, and may change in the future.
}
\examples{
\donttest{
library(terra)
# get predictor variables
f <- list.files(path=paste(system.file(package="phyloregion"),
                '/ex', sep=''), pattern='.tif', full.names=TRUE)

preds <- rast(f)
#plot(preds)
# get species occurrences
d <- read.csv(system.file("ex/Bombax.csv", package="phyloregion"))

# fit ensemble model for four algorithms
mod <- sdmterra(d, predictors = preds, algorithm = "GLM")
}
}
\references{
Zurell, D., Franklin, J., König, C., Bouchet, P.J., Dormann, C.F., Elith, J.,
Fandos, G., Feng, X., Guillera‐Arroita, G., Guisan, A., Lahoz‐Monfort, J.J.,
Leitão, P.J., Park, D.S., Peterson, A.T., Rapacciuolo, G., Schmatz, D.R.,
Schröder, B., Serra‐Diaz, J.M., Thuiller, W., Yates, K.L., Zimmermann, N.E.
and Merow, C. (2020), A standard protocol for reporting species distribution
models. \emph{Ecography}, \strong{43}: 1261-1277.
}
